refactor: 서비스단 인터페이스, 구현체로 분리
